Fixed StandardDBConnector class

+++ src/main/java/controllers/CreateStandardController.java
@@ -43,7 +43,7 @@ public class CreateStandardController {
                 informationAlert.setTitle("The Water");
                 informationAlert.setHeaderText("");
                 informationAlert.showAndWait();
-                StandardDBConnector.saveStandard(Double.parseDouble(temperature.getText()), Integer.parseInt(pH.getText()), Double.parseDouble(dissolvedOxygen.getText()), Double.parseDouble(mlss.getText()));
+                StandardDBConnector.saveStandard(Double.parseDouble(temperature.getText()), Double.parseDouble(pH.getText()), Double.parseDouble(dissolvedOxygen.getText()), Double.parseDouble(mlss.getText()));
                 backToStandard(event);
             } else {
                 Alert errorAlert = new Alert(Alert.AlertType.ERROR,"Could not save because you fill in a form not complete.");
